Stats: Overview reorg and es6-ification
